+++ src/functions/index.ts
@@ -1,7 +1,8 @@
 import { ModuleRegistrationContext } from '@causa/workspace';
 import { EventTopicGenerateCodeForTypeScriptAndJsonEvents } from './event-topic-generate-json.js';
-import { ProjectBuildArtefactForServiceContainer } from './project-build-artefact-service-container.js';
 import { ProjectBuildArtefactForTypeScriptPackage } from './project-build-artefact-ts-package.js';
+import { ProjectBuildArtefactForTypeScriptServerlessFunctions } from './project-build-artefact-ts-serverless-functions.js';
+import { ProjectBuildArtefactForTypeScriptServiceContainer } from './project-build-artefact-ts-service-container.js';
 import { ProjectGetArtefactDestinationForNpmPackage } from './project-get-artefact-destination-npm-package.js';
 import { ProjectPushArtefactForNpmPackage } from './project-push-artefact-npm-package.js';
 import { ProjectReadVersionForJavascript } from './project-read-version-javascript.js';
@@ -9,8 +10,9 @@ import { ProjectReadVersionForJavascript } from './project-read-version-javascri
 export function registerFunctions(context: ModuleRegistrationContext) {
   context.registerFunctionImplementations(
     EventTopicGenerateCodeForTypeScriptAndJsonEvents,
-    ProjectBuildArtefactForServiceContainer,
     ProjectBuildArtefactForTypeScriptPackage,
+    ProjectBuildArtefactForTypeScriptServerlessFunctions,
+    ProjectBuildArtefactForTypeScriptServiceContainer,
     ProjectGetArtefactDestinationForNpmPackage,
     ProjectPushArtefactForNpmPackage,
     ProjectReadVersionForJavascript,